refactor(ai): split AiService into focused services
