+++ backend/src/ai/ai.types.ts
@@ -1,0 +1,21 @@
+export type GeneratedPiece = {
+  type: string;
+  name?: string;
+};
+
+export type ModelCandidate = {
+  provider: string;
+  model: string;
+  client: unknown;
+};
+
+export type AiErrorDetail = {
+  provider: string;
+  model: string;
+  message: string;
+};
+
+export type ProviderModel = {
+  id: string;
+  label: string;
+};
